Set timespan parameters
These aren’t defined in the config API. Instead, the client is
responsible for setting appropriate values.

This feels contrary to HATEOAS. Every client will have to write this.
I’d prefer to see link relations?

Query API is hopefully going to evolve into a more sensible thing as the
performance-datastore sketch develops, so going to live with this for
now.